+++ spec/dummy/config/application.rb
@@ -3,8 +3,6 @@
 require "rails"
 require "action_controller/railtie"
 require "llm_cost_tracker"
-require "llm_cost_tracker/railtie"
-require "llm_cost_tracker/engine"
 
 module Dummy
   class Application < Rails::Application